+++ app/(route)/my/history/[type]/page.tsx
@@ -3,57 +3,11 @@ import theme from '@/styles/theme';
 import { useEffect, useState } from 'react';
 import styled from 'styled-components';
 import dayjs from 'dayjs';
-import { AdapterDayjs } from '@mui/x-date-pickers/AdapterDayjs';
-import { LocalizationProvider } from '@mui/x-date-pickers/LocalizationProvider';
-import { DatePicker } from '@mui/x-date-pickers/DatePicker';
-import Button from '@/components/common/Button';
 import { useRouter } from 'next/navigation';
-import { ProductHistory, useProductHistory } from '@/hooks/queries/useHistory';
-import moment from 'moment';
-const buttonProps = [
-  { children: '최근 2개월', month: 2 },
-  { children: '4개월', month: 4 },
-  { children: '6개월', month: 6 },
-];
-
-const slotStyleProps: any = {
-  textField: {
-    size: 'small',
-    sx: {
-      '& .MuiInputBase-input': {
-        height: '2rem',
-        fontSize: theme.fontSize.body1,
-        fontWeight: 700,
-        padding: '0.5rem',
-      },
-      '& .MuiButtonBase-root': {
-        color: 'black',
-      },
-    },
-  },
-  popper: {
-    sx: {
-      '& .MuiPickersDay-root': {
-        fontSize: theme.fontSize.caption1,
-        fontWeight: 'bold',
-      },
-      '& .MuiTypography-caption ': {
-        fontSize: theme.fontSize.caption1,
-        color: 'black',
-        fontWeight: 'bold',
-        '&:first-of-type': {
-          color: 'red',
-        },
-        '&:last-of-type': {
-          color: 'blue',
-        },
-      },
-      '& .MuiPickersCalendarHeader-labelContainer': {
-        fontSize: theme.fontSize.subtitle3,
-      },
-    },
-  },
-};
+import { useProductHistory } from '@/hooks/queries/useHistory';
+import StateBox from '@/components/History/State';
+import SelectDate from '@/components/History/SelectDate';
+import SetHistoryList from '@/components/History/setHistoryList';
 
 const MyHistory = (props: any) => {
   const [selectState, setSelectState] = useState([1, 0, 0]);
@@ -61,9 +15,6 @@ const MyHistory = (props: any) => {
   const day = dayjs();
   const [endDate, setEndDate] = useState(day);
   const [startDate, setStartDate] = useState(day.add(-2, 'month'));
-  // const [bidding, setBidding] = useState<ProductHistory>();
-  // const [pending, setPending] = useState<ProductHistory>();
-  // const [finished, setFinished] = useState<ProductHistory>();
   const type = props.params.type;
 
   const title = type === 'selling' ? '판매' : '구매';
@@ -91,15 +42,21 @@ const MyHistory = (props: any) => {
   });
 
   const data = [
-    { count: bidding ? bidding?.result.length : 0, title: '입찰' },
-    { count: pending ? pending?.result.length : 0, title: '진행중' },
-    { count: finished ? finished?.result.length : 0, title: '완료' },
-  ];
-  const datePickerValues = [
-    { value: startDate, setValue: setStartDate },
-    { value: endDate, setValue: setEndDate },
+    { count: bidding ? bidding.result.length : 0, title: '입찰' },
+    { count: pending ? pending.result.length : 0, title: '진행중' },
+    { count: finished ? finished.result.length : 0, title: '완료' },
   ];
 
+  const datePickerValues = {
+    start: startDate,
+    end: endDate,
+    setStart: setStartDate,
+    setEnd: setEndDate,
+  };
+
+  const unEasyPick = () => {
+    setSelectEasyPick([0, 0, 0]);
+  };
   const reFetchHandler = () => {
     bidRe();
     pendRe();
@@ -170,79 +127,6 @@ const MyHistory = (props: any) => {
     }
   };
 
-  const setHistoryList = () => {
-    if (selectState[0] === 1) {
-      return (
-        <>
-          {bidding && (
-            <>
-              {bidding.result.map((data, idx) => (
-                <ItemBox key={idx}>
-                  <ProductInfo key={idx}>
-                    <img src={data.productImage} alt={data.productImage} />
-                    <ProductNameOption>
-                      <p id="product_name">{data.productName}</p>
-                      <p id="product_option">{data.size}</p>
-                    </ProductNameOption>
-                  </ProductInfo>
-                  <ItemOption>
-                    <p>{data.price}</p>
-                    <p>{moment(data.deadLine).format('YY-MM-DD')}</p>
-                  </ItemOption>
-                </ItemBox>
-              ))}
-            </>
-          )}
-        </>
-      );
-    } else if (selectState[1] === 1) {
-      return (
-        <>
-          {pending && (
-            <>
-              {pending.result.map((data, idx) => (
-                <ItemBox key={idx}>
-                  <ProductInfo key={idx}>
-                    <img />
-                    <ProductNameOption>
-                      <p id="product_name">{data.productName}</p>
-                      <p id="product_option">{data.size}</p>
-                    </ProductNameOption>
-                  </ProductInfo>
-                  <ItemOption>
-                    <p>{data.status}</p>
-                  </ItemOption>
-                </ItemBox>
-              ))}
-            </>
-          )}
-        </>
-      );
-    } else if (selectState[2] === 1) {
-      return (
-        <>
-          {finished && (
-            <>
-              {finished.result.map((data, idx) => (
-                <ItemBox key={idx}>
-                  <ProductInfo key={idx}>
-                    <img />
-                    <ProductNameOption>
-                      <p id="product_name">{data.productName}</p>
-                      <p id="product_option">{data.size}</p>
-                    </ProductNameOption>
-                  </ProductInfo>
-                  <ItemOption>
-                    <p>{data.status}</p>
-                  </ItemOption>
-                </ItemBox>
-              ))}
-            </>
-          )}
-        </>
-      );
-    }
-  };
   useEffect(() => {
     if (type !== 'selling' && type !== 'buying') {
       alert('로그인 후 이용 가능합니다');
@@ -253,137 +137,34 @@ const MyHistory = (props: any) => {
   return (
     <>
       <h2>{title}내역</h2>
-      <StateWrapper>
-        {data.map((data, idx) => (
-          <State
-            key={idx}
-            onClick={() => stateHandler(idx)}
-            select={selectState[idx]}
-            type={type}
-          >
-            <h3
-              id={'history_count' + (selectState[idx] === 1 ? '_select' : '')}
-            >
-              {data.count ? data.count : 0}
-            </h3>
-            <h4>{title + ' ' + data.title}</h4>
-          </State>
-        ))}
-      </StateWrapper>
-      <SelectDateWrapper>
-        <LocalizationProvider dateAdapter={AdapterDayjs}>
-          <DateWrapper>
-            {buttonProps.map((data, idx) => (
-              <Button
-                size="small"
-                buttonColor="light"
-                styles={{
-                  height: '3rem',
-                  marginTop: '1rem',
-                  marginRight: '1rem',
-                  fontSize: theme.fontSize.caption2,
-                  color:
-                    selectEasyPick[data.month / 2 - 1] === 1
-                      ? 'black'
-                      : theme.colors.gray[200],
-                }}
-                key={idx}
-                onClick={() => dateHandler(data.month)}
-              >
-                {data.children}
-              </Button>
-            ))}
-            {datePickerValues.map((date, idx) => (
-              <DateItem key={idx}>
-                <DatePicker
-                  showDaysOutsideCurrentMonth
-                  value={date.value}
-                  className="date-picker"
-                  format="YYYY-MM-DD"
-                  slotProps={slotStyleProps}
-                  onChange={(newValue) => {
-                    date.setValue(dayjs(newValue));
-                    setSelectEasyPick([0, 0, 0]);
-                  }}
-                />
-              </DateItem>
-            ))}
-            <Button
-              size="small"
-              styles={{
-                height: '4rem',
-                marginTop: '0.55rem',
-                fontSize: theme.fontSize.caption1,
-              }}
-              onClick={reFetchHandler}
-            >
-              조회
-            </Button>
-          </DateWrapper>
-        </LocalizationProvider>
-      </SelectDateWrapper>
+      <StateBox
+        title={title}
+        stateHandler={stateHandler}
+        selectState={selectState}
+        type={type}
+        data={data}
+      />
+      <SelectDate
+        selectEasyPick={selectEasyPick}
+        dateHandler={dateHandler}
+        datePickerValues={datePickerValues}
+        unEasyPick={unEasyPick}
+        reFetchHandler={reFetchHandler}
+      />
       <StateOptionBox>
         <SortOption>{setOption()}</SortOption>
         <OptionTitle>{setTitle()}</OptionTitle>
       </StateOptionBox>
-      {setHistoryList()}
+      <SetHistoryList
+        selectState={selectState}
+        bidding={bidding}
+        pending={pending}
+        finished={finished}
+      />
     </>
   );
 };
 export default MyHistory;
-
-const StateWrapper = styled.div`
-  width: 100%;
-  display: flex;
-  justify-content: center;
-`;
-
-const State = styled.div<{ select: number; type: string }>`
-  width: 30rem;
-  height: 5rem;
-  text-align: center;
-  margin-top: 2.5rem;
-  border-bottom: ${(props) => (props.select === 0 ? '0.1rem' : '0.3rem')} solid
-    ${(props) => (props.select === 0 ? theme.colors.gray[200] : 'black')};
-  cursor: pointer;
-  #history_count_select {
-    color: ${(props) =>
-      props.type === 'selling' ? theme.colors.selling : theme.colors.buying};
-  }
-  #history_count {
-    color: black;
-  }
-  h4 {
-    color: ${(props) =>
-      props.select === 0 ? theme.colors.gray[200] : 'black'};
-  }
-`;
-
-const SelectDateWrapper = styled.div`
-  width: 90rem;
-  height: 10rem;
-  background-color: ${theme.colors.gray[100]};
-  margin: auto;
-  padding-top: 3rem;
-  border-bottom: 0.1rem solid ${theme.colors.gray[200]};
-`;
-
-const DateWrapper = styled.div`
-  display: flex;
-  justify-content: center;
-`;
-
-const DateItem = styled.div`
-  width: 15rem;
-  margin: 1rem;
-  .date-picker {
-    background-color: #fff;
-    border-radius: 0.4rem;
-    div {
-      top: 0.25rem;
-    }
-  }
-`;
 
 const StateOptionBox = styled.div`
   width: 90rem;
@@ -409,44 +190,4 @@ const SortOption = styled.select`
   width: 12rem;
   height: 4rem;
   cursor: pointer;
-`;
-
-const ItemBox = styled.div`
-  width: 90rem;
-  margin: 1rem auto;
-  padding: 1rem;
-  display: flex;
-  justify-content: space-between;
-  border: 0.1rem solid ${theme.colors.gray[200]};
-  border-radius: 1rem;
-  img {
-    width: 8rem;
-    height: 8rem;
-    float: left;
-    margin-right: 1rem;
-  }
-`;
-
-const ProductInfo = styled.div`
-  display: flex;
-`;
-
-const ItemOption = styled.div`
-  display: flex;
-  line-height: 8rem;
-  p {
-    margin: 0 5rem;
-  }
-`;
-
-const ProductNameOption = styled.div`
-  padding: 1rem 0 0 1rem;
-  #product_name {
-    font-weight: bold;
-    margin-bottom: 0.5rem;
-  }
-  #product_option {
-    font-size: ${theme.fontSize.subtitle3};
-    color: ${theme.colors.gray[200]};
-  }
 `;
